requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
aiofiles
python-multipart
supabase-py